Use Bootstrap CSS and set up site structure accordingly.
This allows us to mimic the structure of the existing site as closely as possible,
while keeping it in line with any future changes to the base code. This isn't my
preferred method, but I think it makes sense in this case.

Note that the sidebar has been moved before the content, since that's how the original
site was structured. It may be wise to move it back, and move the content on larger screens
with CSS, but that will likely depend on what's specifically required here.